pyproject.toml
[tool.poetry]
name = "board"
version = "0.1.0"
description = ""
authors = ["me"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
drf-yasg = "^1.21.8"
django-cors-middleware = "^1.5.0"
django-phonenumber-field = "^8.0.0"
phonenumbers = "^8.13.52"
pillow = "^11.0.0"
djoser = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
